drivers: rtc: pcf85263: fix incompatible-pointer-types for GCC 14+
GCC 14 makes -Wincompatible-pointer-types an error. pcf85263_update_bits()
read the register into a uint32_t local and passed its address to
pcf85263_read(), which takes a uint8_t*. Declare the local as uint8_t to
match the accessor; the subsequent mask/merge and pcf85263_write() call
both operate on uint8_t, so the value is unchanged.

+++ drivers/rtc/pcf85263/pcf85263.c
@@ -84,7 +84,7 @@ int pcf85263_update_bits(struct pcf85263_dev *dev, uint8_t reg_addr,
 			 uint8_t mask, uint8_t reg_data)
 {
 	int ret;
-	uint32_t data;
+	uint8_t data;
 
 	ret = pcf85263_read(dev, reg_addr, &data);
 	if (ret)
